SwaggerUI: Separated logic for middleware for re-use
