+++ src/Swashbuckle.AspNetCore.SwaggerUI/SwaggerUIMiddleware.cs
@@ -3,8 +3,6 @@ using System.Threading.Tasks;
 using System.Text.RegularExpressions;
 using System.Linq;
 using System.Text;
-using System.Collections.Generic;
-using System.IO;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Http;
@@ -12,9 +10,6 @@ using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 using Microsoft.Extensions.FileProviders;
 using Microsoft.AspNetCore.StaticFiles;
-using Newtonsoft.Json;
-using Newtonsoft.Json.Serialization;
-using Newtonsoft.Json.Converters;
 #if NETCOREAPP3_0
 using IHostingEnvironment = Microsoft.AspNetCore.Hosting.IWebHostEnvironment;
 #endif
@@ -26,8 +21,8 @@ namespace Swashbuckle.AspNetCore.SwaggerUI
         private const string EmbeddedFileNamespace = "Swashbuckle.AspNetCore.SwaggerUI.node_modules.swagger_ui_dist";
 
         private readonly SwaggerUIOptions _options;
-        private readonly JsonSerializer _jsonSerializer;
         private readonly StaticFileMiddleware _staticFileMiddleware;
+        private readonly SwaggerUIIndexHtmlBuilder _swaggerUIIndexHtmlBuilder;
 
         public SwaggerUIMiddleware(
             RequestDelegate next,
@@ -44,8 +39,8 @@ namespace Swashbuckle.AspNetCore.SwaggerUI
             SwaggerUIOptions options)
         {
             _options = options ?? new SwaggerUIOptions();
-            _jsonSerializer = CreateJsonSerializer();
             _staticFileMiddleware = CreateStaticFileMiddleware(next, hostingEnv, loggerFactory, options);
+            _swaggerUIIndexHtmlBuilder = new SwaggerUIIndexHtmlBuilder(_options);
         }
 
         public async Task Invoke(HttpContext httpContext)
@@ -74,18 +69,6 @@ namespace Swashbuckle.AspNetCore.SwaggerUI
             await _staticFileMiddleware.Invoke(httpContext);
         }
 
-        private JsonSerializer CreateJsonSerializer()
-        {
-            return JsonSerializer.Create(new JsonSerializerSettings
-            {
-                ContractResolver = new CamelCasePropertyNamesContractResolver(),
-                Converters = new[] { new StringEnumConverter(true) },
-                NullValueHandling = NullValueHandling.Ignore,
-                Formatting = Formatting.None,
-                StringEscapeHandling = StringEscapeHandling.EscapeHtml
-            });
-        }
-
         private StaticFileMiddleware CreateStaticFileMiddleware(
             RequestDelegate next,
             IHostingEnvironment hostingEnv,
@@ -112,35 +95,9 @@ namespace Swashbuckle.AspNetCore.SwaggerUI
             response.StatusCode = 200;
             response.ContentType = "text/html;charset=utf-8";
 
-            using (var stream = _options.IndexStream())
-            {
-                // Inject arguments before writing to response
-                var htmlBuilder = new StringBuilder(new StreamReader(stream).ReadToEnd());
-                foreach (var entry in GetIndexArguments())
-                {
-                    htmlBuilder.Replace(entry.Key, entry.Value);
-                }
+            var indexHtml = _swaggerUIIndexHtmlBuilder.Build();
 
-                await response.WriteAsync(htmlBuilder.ToString(), Encoding.UTF8);
-            }
-        }
-
-        private IDictionary<string, string> GetIndexArguments()
-        {
-            return new Dictionary<string, string>()
-            {
-                { "%(DocumentTitle)", _options.DocumentTitle },
-                { "%(HeadContent)", _options.HeadContent },
-                { "%(ConfigObject)", SerializeToJson(_options.ConfigObject) },
-                { "%(OAuthConfigObject)", SerializeToJson(_options.OAuthConfigObject) }
-            };
-        }
-
-        private string SerializeToJson(object obj)
-        {
-            var writer = new StringWriter();
-            _jsonSerializer.Serialize(writer, obj);
-            return writer.ToString();
-        }
+            await response.WriteAsync(indexHtml, Encoding.UTF8);
+        } 
     }
 }
